+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-07-30
+### Added 
+- Visit model
+- Model#insurances helper for Patient.Insurances || Visit.Insurances
+
 ## [1.1.1] - 2020-06-15
 ### Changed
 - Bugfix with patient param in update
@@ -98,6 +103,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial Release
 
+[1.2.0]: https://github.com/WeInfuse/redox/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/WeInfuse/redox/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/WeInfuse/redox/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/WeInfuse/redox/compare/v1.0.1...v1.0.2